+++ CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [13.7.0] - 2023-09-18
+## [Unreleased]
 
 ## Added
-- `sendOrderWithoutSecretKey` method to `MultisigWallet`
+- `TonClient4.getAccountTransactionsParsed` method (thanks @vzhovnitsky)
+- `MultisigWallet` contract (thanks @Gusarich)
+- `WalletV5` contract (thanks @siandreev)
+- blockchain fees estimation via `computeStorageFees`/`computeExternalMessageFees`/`computeGasPrices`/`computeMessageForwardFees` (thanks @vzhovnitsky)
 
 ## Fixed
-- Uri encode get method name for `TonClient4`
+- Uri encode get method name for `TonClient4` (thanks @krigga)
+- Improved `parseStackItem` due to toncenter.com bug
 
 
 ## [13.6.1] - 2023-08-24